Profile edit bootstrap rating

+++ protected/controllers/UserController.php
@@ -33,11 +33,11 @@ class UserController extends Controller
 			$model->location = CHtml::encode($_POST['location']);
 			$model->homepage = CHtml::encode($_POST['homepage']);
 			$model->description = CHtml::encode($_POST['description']);
+			$model->invitations_left = CHtml::encode($_POST['invitations_left']);
             
-            $uploadCover='';
-            if(isset($_POST['UserData']['extension'])){
-			    //Cover Datei
-			    $uploadCover = CUploadedFile::getInstance($model,'extension');
+		    //Cover Datei
+		    $uploadCover = CUploadedFile::getInstance($model,'extension');
+            if($uploadCover){
                 $p = $uploadCover;
 			    $covername = "{$uploadCover}";
 			    $coverInfo = pathinfo($covername);
@@ -51,16 +51,10 @@ class UserController extends Controller
 			    //Dateien Speichern
 			    if($uploadCover){
 				    $uploadCover->saveAs(Yii::app()->basePath.'/../upload/user/original/'.$model->id.'.'.$model->extension);
-                    try{
+                    if(is_file(Yii::getPathOfAlias('webroot').'/upload/user/comment/'.$model->id.'.'.$model->extension))
                         unlink(Yii::getPathOfAlias('webroot').'/upload/user/comment/'.$model->id.'.'.$model->extension);
-			        }
-                    catch(Exception $e)
-                    { }
-                    try{
+                    if(is_file(Yii::getPathOfAlias('webroot').'/upload/user/big/'.$model->id.'.'.$model->extension))
                         unlink(Yii::getPathOfAlias('webroot').'/upload/user/big/'.$model->id.'.'.$model->extension);
-			        }
-                    catch(Exception $e)
-                    { }
 		        }
 		        $this->redirect(array('user/profile','id'=>YII::app()->user->id));
 		    }
